refactor(demo): strip auth and team-specific features
Phase 2.2 — remove login, auth callback, protected routes, and
auth-dependent settings from the demo app. The demo runs in
demo/sandbox mode only and doesn't need authentication.
