--- apps/demo/src/components/protected-route.tsx
@@ -1,49 +1,0 @@
-import { Navigate, useLocation } from "@tanstack/react-router";
-
-import { useAuth } from "../contexts";
-
-interface ProtectedRouteProps {
-  children: React.ReactNode;
-  requiredRole?: "admin" | "operator" | "viewer";
-}
-
-export function ProtectedRoute({ children, requiredRole }: ProtectedRouteProps) {
-  const { isAuthenticated, isLoading, user } = useAuth();
-  const location = useLocation();
-
-  // Show loading state while checking auth
-  if (isLoading) {
-    return (
-      <div className="min-h-screen flex items-center justify-center bg-background">
-        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-primary" />
-      </div>
-    );
-  }
-
-  // Redirect to login if not authenticated
-  if (!isAuthenticated) {
-    return <Navigate to="/login" replace />;
-  }
-
-  // Check role if required
-  if (requiredRole && user) {
-    const roleHierarchy = { viewer: 0, operator: 1, admin: 2 };
-    const userRoleLevel = roleHierarchy[user.role] ?? 0;
-    const requiredRoleLevel = roleHierarchy[requiredRole] ?? 0;
-
-    if (userRoleLevel < requiredRoleLevel) {
-      return (
-        <div className="min-h-screen flex items-center justify-center bg-background">
-          <div className="text-center">
-            <h1 className="text-2xl font-bold text-destructive mb-2">Access Denied</h1>
-            <p className="text-muted-foreground">
-              You don't have permission to access this page.
-            </p>
-          </div>
-        </div>
-      );
-    }
-  }
-
-  return <>{children}</>;
-}
