+++ apps/demo/src/components/settings/index.ts
@@ -1,5 +1,4 @@
 export { ProfileSettings } from "./profile-settings";
-export { SecuritySettings } from "./security-settings";
 export { ApiKeySettings } from "./api-key-settings";
 export { OrgSettings } from "./org-settings";
 export { AppearanceSettings } from "./appearance-settings";
