+++ apps/demo/src/contexts/index.ts
@@ -1,2 +1,1 @@
-export { AuthProvider, useAuth, useOAuthCallback, type User } from "./auth-context";
 export { SidePanelProvider, useSidePanel } from "./side-panel-context";
