+++ apps/demo/src/pages/index.ts
@@ -1,9 +1,7 @@
 export * from "./agents";
-export * from "./auth-callback";
 export * from "./credits";
 export * from "./dashboard";
 export * from "./events";
-export * from "./login";
 export * from "./messages";
 export * from "./settings";
 export * from "./tasks";
